rag_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
qdrant-client==1.11.1
sentence-transformers==3.0.1
httpx==0.27.2
sse-starlette==2.1.2
jinja2==3.1.4
pypdf==4.3.1
pyyaml==6.0.2
